+++ framework/src/Volo.Abp.Cli.Core/Volo/Abp/ProjectModification/SolutionModuleAdder.cs
@@ -1,12 +1,16 @@
 ﻿using System.Threading.Tasks;
+using JetBrains.Annotations;
 using Volo.Abp.DependencyInjection;
 
 namespace Volo.Abp.ProjectModification
 {
     public class SolutionModuleAdder : ITransientDependency
     {
-        public async Task AddAsync(string solutionFile, string moduleName)
+        public virtual async Task AddAsync([NotNull] string solutionFile, [NotNull] string moduleName)
         {
+            Check.NotNull(solutionFile, nameof(solutionFile));
+            Check.NotNull(moduleName, nameof(moduleName));
+
 
         }
     }
